workflows: adapt for aviant 1.15 branch
Squash and minor adaptation of cd8b050, e9a204e and 414f987 from the 1.13 branch

Remove workflwes and steps/builds not relevant. Remove uploads to
services that we do not use. Adapt for branch aviant/1.15 instead of
master and use for PR for all branch names (would not run for branches
with "/" in the name).

Only build default version in workflow
The workflow would build cubepilot_cubeorange_test and _bootloader, that
we do not use. And cubepilot_cubeorange_test is using 100% flash, and
will fail.

workflow: releases built on tag push will be pre-release
This is normally what we want, and we will manually update it to
"latest" when it is ready for production.