Return to center after loiter to altitude
This commit revert most of
https://github.com/PX4/PX4-Autopilot/pull/18996
so that we will no longer skip ahead to next waypoint if we are ahead of
the current. This will enfource that we return to the center of the
loiter before continue after a loiter to altitude. This is to prevent
that we turn out of the loiter when done, and reach a distance from the
WP/path that is larger than intended.