ekf2: allow separate instances for VO/flow and GNSS
This is controlled by the parameter EFK2_GNSS_DENIED.
VO/flow instances are only for logging, will not be selected for navigation.
As of now, the GNSS-denied instances will fuse GNSS while on the ground
for initialization. This is controlled by param EKF2_GD_GPS_INIT

To avoid making too many instances, make sure to use SENS_MAG_MODE=1