boards/cubepilot/cubeorange: disable one internal imu
When running multi-EKF, individual IMUs cannot be disabled through parameters.
We want to run multi-EKF with two internal IMUs and one with external vectornav
IMU.

In order to achieve this, we disable one of the internal IMU by simply not
starting the driver.

We disable from SPI4, where there are two, so that we have redundancy on SPI1.
We leave the icm20948, so that we still have magnetometer