ekf2: don't use EKF2_REQ_FIXTYPE after initializing origin
Previously, EKF2_REQ_FIXTYPE was used for all GPS checks.
This could lead to a situation where GPS drops out and then is regained, but the measurements are still rejected due to failing checks. In this case, we want any measurement with fix type >=3 to be accepted.

This commit reverts to regular fix 3 requirement after the EKF has been initialized, but keeps EKF2_REQ_FIXTYPE requirement for initialization.