sensors: add vehicle_gnss_heading to sensor bridge
1.15 port of 8901aba

Modelled after vehicle_gps_position, it processes incoming sensor_gnss_relative
messages and outputs vehicle_gps_yaw with a validated body-frame heading
measurement.

In the future, this can be used in the EKF, instead of the existing hacky
solution which only works with blending enabled.

Note that only one instance is supported. This is because PX4 only has support
for two GNSS receivers anyway, so there isn't "room" for multiple rovers that
output heading.