gps: option for normal mode fallback if ubx mb rover is used for position
A receiver in moving base rover mode will output data with delay up to 1 second
if it cannot compute a relative solution. This can for example happen if the
moving base experiences degrated signal or power loss.

In such a scenario, the rover might be the only receiver left, and for
redundancy it's important to bring it back into a normal state without delay in
order to avoid toilet bowl/position loss.